+++ vitrina/uapi/urls.py
@@ -1,5 +1,8 @@
 from django.urls import path
 
+from vitrina.uapi.views.agreement_file_views import AgreementFileDownloadUAPIView
+from vitrina.uapi.views.agreement_views import AgreementViewSet
+
 from vitrina.uapi.views.template_views import (
     AgentDeleteView,
     AgentUpdateView,
@@ -73,6 +76,11 @@ urlpatterns = [
         name="uapi-distribution",
     ),
     path(
+        f"{STATIC_UAPI_BASE_PATH}Agreement/",
+        AgreementViewSet.as_view({"get": "list"}),
+        name="uapi-agreement",
+    ),
+    path(
         f"{STATIC_UAPI_BASE_PATH}Agreement/<uuid:agreement_id>/sync-done/",
         AgentSyncDoneViewSet.as_view({"put": "update"}),
         name="uapi-agent-sync-done",
@@ -86,5 +94,10 @@ urlpatterns = [
         f"{STATIC_UAPI_BASE_PATH}Agent/",
         AgentViewSet.as_view({"get": "list"}),
         name="uapi-agent",
+    ),
+    path(
+        f"{STATIC_UAPI_BASE_PATH}AgreementFile/<uuid:agreement_file_uuid>/file/",
+        AgreementFileDownloadUAPIView.as_view(),
+        name="uapi-agreement-file-download",
     ),
 ]